# web/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Marketing Api",
        "version": "1.0.0"
    },
    "host": "api5.audi.dev.agence-one.net",
    "basePath": "/marketing/dev/",
    "schemes": [
        "http"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/invitation/{id}": {
            "get": {
                "tags": [
                    "Invitation"
                ],
                "summary": "Get invitation",
                "operationId": "getInvitation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Invitation ID",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Invitation",
                        "schema": {
                            "$ref": "#/definitions/BasicEntityContext"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Invitation"
                ],
                "summary": "update invitation from id",
                "operationId": "putInvitationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "invitationId",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Invitation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Invitation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the updated invitation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/InvitationComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Invitation page not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Invitation"
                ],
                "summary": "remove invitation from id",
                "operationId": "removeInvitationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "invitationId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the invitation"
                    },
                    "404": {
                        "description": "not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Invitation"
                ],
                "summary": "patch invitation from id",
                "operationId": "patchInvitationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "invitationId",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "invitation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Invitation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the updated invitation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/InvitationComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Invitation page not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/invitation": {
            "get": {
                "tags": [
                    "Invitation"
                ],
                "summary": "Get Invitations",
                "operationId": "getInvitations",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Get invitations where status",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitations",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Invitation"
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/invitation/": {
            "post": {
                "tags": [
                    "Invitation"
                ],
                "summary": "create invitation",
                "operationId": "createInvitation",
                "parameters": [
                    {
                        "name": "invitation",
                        "in": "body",
                        "description": "Invitation payload",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Invitation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "the created invitation",
                        "schema": {
                            "$ref": "#/definitions/InvitationComplete"
                        }
                    },
                    "404": {
                        "description": "Invitation page not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/invitation-user": {
            "get": {
                "tags": [
                    "Invitation"
                ],
                "summary": "Get InvitationsUser",
                "operationId": "getInvitationsUser",
                "parameters": [
                    {
                        "name": "invitationId",
                        "in": "query",
                        "description": "Get myAudiUserInvitations by invitationId",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "Get myAudiUserInvitations by userId",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "InvitationsUser",
                        "schema": {
                            "$ref": "#/definitions/MyaudiUserInvitationContextMulti"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/invitation-user/": {
            "post": {
                "tags": [
                    "Invitation"
                ],
                "summary": "Subscribe a user to an invitation",
                "operationId": "postMyaudiUserInvitation",
                "parameters": [
                    {
                        "name": "myaudiUserInvitationPayload",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MyaudiUserInvitationPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription",
                        "schema": {
                            "$ref": "#/definitions/MyaudiUserInvitation"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/invitation-user/{id}": {
            "delete": {
                "tags": [
                    "Invitation"
                ],
                "summary": "Delete MyaudiUserInvitation by id",
                "operationId": "deleteMyaudiUserInvitationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "MyaudiUserInvitation Id to delete",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete status",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "Context": {
            "properties": {
                "context": {
                    "type": "string"
                }
            },
            "type": "object"
        },
        "Success": {
            "properties": {
                "data": {
                    "properties": {
                        "message": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/Context"
                }
            ]
        },
        "Error": {
            "required": [
                "errors"
            ],
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "message": {
                                "type": "string"
                            },
                            "code": {
                                "type": "integer"
                            },
                            "type": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/Context"
                }
            ]
        },
...
        "TimestampableEntity": {
            "properties": {
                "createdAt": {
                    "type": "string",
                    "format": "date-time",
                    "default": "now"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time",
                    "default": "now"
                }
            }
        },
...
        "BasicEntityContext": {
            "properties": {
                "data": {
                    "type": "object",
                    "allOf": [
                        {
                            "$ref": "#definitions/BasicEntity"
                        }
                    ]
                }
            },
            "allOf": [
                {
                    "$ref": "#definitions/Context"
                }
            ]
        },
...
        "Invitation": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "campaignEvent": {
                    "type": "object",
                    "allOf": [
                        {
                            "$ref": "#definitions/CampaignEvent"
                        }
                    ]
                },
                "description": {
                    "type": "string"
                },
                "teaser": {
                    "type": "string"
                },
                "mailto": {
                    "type": "string"
                },
                "pathDesktop": {
                    "type": "string"
                },
                "pathTablet": {
                    "type": "string"
                },
                "pathMobile": {
                    "type": "string"
                }
            },
            "allOf": [
                {
                    "$ref": "#definitions/BasicEntity"
                }
            ]
        },
        "MyaudiUserInvitation": {
            "properties": {
                "myaudiUserId": {
                    "type": "integer"
                },
                "marketingObjective": {
                    "type": "object",
                    "allOf": [
                        {
                            "$ref": "#definitions/MarketingObjective"
                        }
                    ]
                },
                "isUnsubscribe": {
                    "type": "boolean"
                }
            }
        },
        "MyaudiUserInvitationPayload": {
            "properties": {
                "myaudiUserId": {
                    "type": "integer"
                },
                "invitation": {
                    "type": "integer"
                }
            }
        },
        "MyaudiUserInvitationContextMulti": {
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#definitions/MyaudiUserInvitation"
                    }
                }
            },
            "allOf": [
                {
                    "$ref": "#definitions/Context"
                }
            ]
        },
        "InvitationComplete": {
            "allOf": [
                {
                    "$ref": "#/definitions/TimestampableEntity"
                }
            ]
        }
    }
}